Moves editions page tests into separate file
